Fix api security problems

+++ src/main/java/eu/cryptoeuro/accountmapper/rest/AccountMapperController.java
@@ -33,6 +33,7 @@ import java.security.Principal;
 import java.util.Base64;
 import java.util.List;
 
+import static java.util.Collections.emptyList;
 import static org.springframework.web.bind.annotation.RequestMethod.*;
 
 @RestController
@@ -205,12 +206,8 @@ public class AccountMapperController {
 																										  @RequestHeader(value = "Authorization") String authorization) {
 		AccountActivationResponse.AccountActivationResponseBuilder responseBuilder = AccountActivationResponse.getBuilderForAuthType(AuthorisationType.BANK_TRANSFER).ownerId(cmd.getOwnerId());
 
-		//Wow, such authorization, much secure
-		try {
-			if (authorization == null || !ethereumService.getParityAuthCredentials().equals(new String(Base64.getDecoder().decode(authorization.replace("Basic ", "")))))
-				throw new Exception("Account registration via bank unauthorized");
-		} catch (Exception e) {
-			log.error(e.getLocalizedMessage());
+		if (!isAuthorized(authorization)) {
+			log.error("Account registration via bank unauthorized");
 			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 		}
 
@@ -248,20 +245,37 @@ public class AccountMapperController {
 			@RequestParam(name = "inactive", required = false) boolean inactive
 	) {
 		if (ownerId != null) {
-			return new ResponseEntity<AccountsResponse>(AccountsResponse.fromEthereumAccounts(accountManagementService.getAccountsByOwnerIdActiveEscrow(ownerId,inactive,escrow)), HttpStatus.OK);
+			return new ResponseEntity<>(AccountsResponse.fromEthereumAccounts(accountManagementService.getAccountsByOwnerIdActiveEscrow(ownerId,inactive,escrow)), HttpStatus.OK);
 		}
-		return new ResponseEntity<AccountsResponse>(AccountsResponse.fromEthereumAccounts(accountManagementService.getAllAccounts()), HttpStatus.OK);
+		return new ResponseEntity<>(AccountsResponse.fromEthereumAccounts(emptyList()), HttpStatus.BAD_REQUEST);
 	}
 
 	@ApiOperation(value = "Remove account identity mapping")
 	@RequestMapping(method = DELETE, value = "/accounts")
-	public ResponseEntity<AccountsResponse> removeAccount(@RequestParam(name = "mappingId", required = true) Long mappingId) {
+	public ResponseEntity<AccountsResponse> removeAccount(
+			@RequestParam(name = "mappingId", required = true) Long mappingId,
+			@RequestHeader(value = "Authorization") String authorization) {
+
+		if (!isAuthorized(authorization)) {
+			log.error("Account deletion unauthorized");
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
 		accountManagementService.removeAccountById(mappingId);
-		return new ResponseEntity<AccountsResponse>(AccountsResponse.fromEthereumAccounts(accountManagementService.getAllAccounts()), HttpStatus.OK);
+		return new ResponseEntity<>(AccountsResponse.fromEthereumAccounts(accountManagementService.getAllAccounts()), HttpStatus.OK);
 	}
 
+    private boolean isAuthorized(String authorization) {
+        try {
+            return authorization != null && ethereumService.getParityAuthCredentials().trim().equals(new String(Base64.getDecoder().decode(authorization.replace("Basic ", ""))));
+        } catch (IOException e) {
+            log.error("Exception checking authorization", e);
+            return false;
+        }
+    }
+
 	private List<EscrowTransfer> clearEscrow(EthereumAccount account) throws IOException,JSONException {
-			
+
 	  	long idCode = Long.parseLong(account.getOwnerId());
 		String address  = account.getAddress();
 		// check if escrow
@@ -270,9 +284,4 @@ public class AccountMapperController {
 			return escrowService.clearAllToAddress(idCode,address);
 		} else return null;
 	}
-
-
-
-
-
 }
